Implemented stats service and use the catalogue method to record stats

+++ lobcder/src/main/java/nl/uva/cs/lobcder/rest/wrappers/Stats.java
@@ -25,9 +25,8 @@ import lombok.Data;
 @XmlRootElement
 @Data
 public class Stats {
-
     String source;
     String destination;
-    Double size;
+    Long size;
     Double speed;
 }
